requirements.txt
crewai==0.114.0
tavily-python
scrapegraph-py
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
agentops
